tmfront/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pytesseract
pillow
python-jose
passlib[bcrypt]==1.7.4
email-validator
python-multipart
openpyxl
bcrypt
